refactor: Address PR feedback - extract shared utilities and fix duplication
- Create poc/utils.py with shared utilities:
  - SUPPORTED_AUDIO_FORMATS for consistent audio format handling
  - format_timestamp() - shared timestamp formatting
  - extract_audio_segment() - shared audio extraction (was duplicated in 4 files)
  - resolve_song_audio_path() - shared song ID resolution (was duplicated in 5 files)

- Fix gen_clean_vocal_stem.py: Use _find_stage1_stems() helper instead of duplicated logic

- Fix gen_lrc_omnisensevoice.py: Fix audio format inconsistency (.wav/.mp3 vs .flac/.m4a)

- Update all gen_lrc_*.py scripts to import from poc/utils.py:
  - gen_lrc_sensevoice.py
  - gen_lrc_whisper.py
  - gen_lrc_whisperx.py
  - gen_lrc_qwen3.py (uses shared format_timestamp, keeps custom resolution for offline mode)

Code reduction: ~616 lines removed
