+++ poc/gen_lrc_omnisensevoice.py
@@ -5,66 +5,15 @@ Runs OmniSenseVoice transcription on a song from the local cache or a direct
 file path, enabling quick experimentation to diagnose transcription accuracy.
 """
 
-import sys
 from pathlib import Path
 from typing import Optional
 
 import typer
 
-# Add src to path for imports
-sys.path.insert(0, str(Path(__file__).parent.parent / "src"))
-
-from stream_of_worship.app.config import AppConfig
-from stream_of_worship.app.db.read_client import ReadOnlyClient
-from stream_of_worship.app.services.catalog import CatalogService
-from stream_of_worship.app.services.asset_cache import AssetCache
-from stream_of_worship.admin.services.r2 import R2Client
+# Import shared utilities
+from poc.utils import SUPPORTED_AUDIO_FORMATS, extract_audio_segment, format_timestamp
 
 app = typer.Typer(help="OmniSenseVoice transcription driver")
-
-
-def format_timestamp(seconds: float) -> str:
-    """Format seconds as [mm:ss.xx] timestamp."""
-    minutes = int(seconds // 60)
-    secs = seconds % 60
-    return f"[{minutes:02d}:{secs:05.2f}]"
-
-
-def extract_audio_segment(
-    audio_path: Path,
-    start_seconds: float,
-    end_seconds: float,
-) -> Path:
-    """Extract a segment of audio to a temporary file."""
-    from pydub import AudioSegment
-
-    audio = AudioSegment.from_file(str(audio_path))
-
-    duration_ms = len(audio)
-    start_ms = int(start_seconds * 1000)
-    end_ms = int(end_seconds * 1000)
-
-    start_ms = max(0, start_ms)
-    end_ms = min(duration_ms, end_ms)
-
-    if start_ms >= end_ms:
-        raise ValueError(f"Invalid segment: start ({start_seconds}s) >= end ({end_seconds}s)")
-
-    segment = audio[start_ms:end_ms]
-
-    import tempfile
-
-    source_suffix = audio_path.suffix.lower()
-    if source_suffix not in (".wav", ".mp3"):
-        source_suffix = ".wav"
-
-    temp_file = tempfile.NamedTemporaryFile(suffix=source_suffix, delete=False)
-    temp_path = Path(temp_file.name)
-    temp_file.close()
-
-    export_format = source_suffix.lstrip(".")
-    segment.export(str(temp_path), format=export_format)
-    return temp_path
 
 
 def transcribe_audio(
@@ -147,7 +96,7 @@ def transcribe_audio(
             import tempfile
 
             source_suffix = transcribe_path.suffix.lower()
-            if source_suffix not in (".wav", ".mp3", ".flac", ".m4a"):
+            if source_suffix not in SUPPORTED_AUDIO_FORMATS:
                 source_suffix = ".wav"
 
             audio = AudioSegment.from_file(str(transcribe_path))
@@ -312,85 +261,13 @@ def main(
     ),
 ):
     """Run OmniSenseVoice transcription on a song and output LRC format."""
+    from poc.utils import resolve_song_audio_path
+
     if textnorm not in ("withitn", "woitn"):
         typer.echo("Error: --textnorm must be one of: withitn, woitn", err=True)
         raise typer.Exit(1)
 
-    input_path = Path(song_id).expanduser()
-    audio_path: Optional[Path] = None
-    if input_path.exists():
-        audio_path = input_path
-        typer.echo(f"Using direct audio path: {audio_path}", err=True)
-    else:
-        try:
-            config = AppConfig.load()
-        except FileNotFoundError:
-            typer.echo(
-                "Error: Config file not found. Please run 'sow-app' first to create config.",
-                err=True,
-            )
-            raise typer.Exit(1)
-
-        db_client = ReadOnlyClient(config.db_path)
-        catalog = CatalogService(db_client)
-
-        song_with_recording = catalog.get_song_with_recording(song_id)
-        if not song_with_recording:
-            typer.echo(f"Error: Song not found: {song_id}", err=True)
-            raise typer.Exit(1)
-
-        if not song_with_recording.recording:
-            typer.echo(f"Error: No recording found for song: {song_id}", err=True)
-            raise typer.Exit(1)
-
-        song = song_with_recording.song
-        recording = song_with_recording.recording
-        hash_prefix = recording.hash_prefix
-
-        typer.echo(f"Song: {song.title}", err=True)
-        typer.echo(f"Recording: {hash_prefix}", err=True)
-
-        try:
-            r2_client = R2Client(
-                bucket=config.r2_bucket,
-                endpoint_url=config.r2_endpoint_url,
-                region=config.r2_region,
-            )
-        except ValueError as e:
-            typer.echo(f"Error: R2 credentials not configured: {e}", err=True)
-            raise typer.Exit(1)
-
-        cache = AssetCache(cache_dir=config.cache_dir, r2_client=r2_client)
-
-        if use_vocals:
-            vocals_path = cache.get_stem_path(hash_prefix, "vocals")
-            if vocals_path.exists():
-                audio_path = vocals_path
-                typer.echo(f"Using cached vocals stem: {audio_path}", err=True)
-            else:
-                typer.echo("Downloading vocals stem...", err=True)
-                audio_path = cache.download_stem(hash_prefix, "vocals")
-                if audio_path:
-                    typer.echo(f"Downloaded vocals stem: {audio_path}", err=True)
-
-        if audio_path is None:
-            main_audio_path = cache.get_audio_path(hash_prefix)
-            if main_audio_path.exists():
-                audio_path = main_audio_path
-                typer.echo(f"Using cached main audio: {audio_path}", err=True)
-            else:
-                typer.echo("Downloading main audio...", err=True)
-                audio_path = cache.download_audio(hash_prefix)
-                if audio_path:
-                    typer.echo(f"Downloaded main audio: {audio_path}", err=True)
-
-        if audio_path is None:
-            typer.echo("Error: Could not find or download audio", err=True)
-            raise typer.Exit(1)
-
-        if not audio_path.exists():
-            typer.echo(f"Error: Audio file not found: {audio_path}", err=True)
-            raise typer.Exit(1)
+    audio_path, _ = resolve_song_audio_path(song_id, use_vocals=use_vocals)
 
     effective_end: Optional[float] = end if end and end > 0 else None
     if effective_end:
